Add launch wrapper to Calamares to fix GNOME starting issue
On GNOME-Shell, pkexec ends up without a parent, resulting in a failure
to start the actual application.
By wrapping the pkexec calamares command in a shell-script, we fix this
by creating a new parent.
I currently consider this a workaround, so it is only applied to tanglu-
live-installer.desktop.
If it turns out that GNOME won't fix this issue, I will submit a patch
for calamares.desktop to Calamares upstream.